Change of data loading logic
Data is now loaded at the start of the app and stored in session_state
